--- docs/source/yirgacheffe.backends.rst
@@ -1,37 +1,0 @@
-yirgacheffe.backends package
-============================
-
-Submodules
-----------
-
-yirgacheffe.backends.enumeration module
----------------------------------------
-
-.. automodule:: yirgacheffe.backends.enumeration
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-yirgacheffe.backends.mlx module
--------------------------------
-
-.. automodule:: yirgacheffe.backends.mlx
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-yirgacheffe.backends.numpy module
----------------------------------
-
-.. automodule:: yirgacheffe.backends.numpy
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Module contents
----------------
-
-.. automodule:: yirgacheffe.backends
-   :members:
-   :show-inheritance:
-   :undoc-members:
