+++ tests/test_auto_windowing.py
@@ -251,7 +251,7 @@ def test_vector_layers_multiply() -> None:
         expected = np.array([[2, 0], [0, 8]])
         assert (expected == actual).all()
 
-@pytest.mark.skipif(yirgacheffe.backends.BACKEND != "NUMPY", reason="Only applies for numpy")
+@pytest.mark.skipif(yirgacheffe._backends.BACKEND != "NUMPY", reason="Only applies for numpy")
 def test_parallel_save_windows() -> None:
     data1 = np.array([[1, 2], [3, 4]])
     data2 = np.array([[10, 20, 30, 40], [50, 60, 70, 80], [90, 100, 110, 120], [130, 140, 150, 160]])
